pyproject.toml
[tool.poetry]
name = "bank-app"
version = "0.1.0"
description = ""
authors = ["Egor Nikitin <eganik5674@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
pandas = "^2.1.3"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.2"
xlrd = "^2.0.1"
pandas-stubs = "^2.1.1.230928"
types-requests = "^2.31.0.10"


[tool.poetry.group.lint.dependencies]
flake8 = "^6.1.0"
mypy = "^1.7.1"
isort = "^5.12.0"
black = "^23.11.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    )/

)
'''

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
